Story/q20 8 (#10)
* [Q20-8] network endpoint

* [Q20-8] network data model

* [Q20-8] mongo Impl and json Parse

* [Q20-8] resolve codefactor issue

* Changed the network object according to the graph JSON, added default
mapping.

* Returning response status after adding the network

* Changed the network class

* Added changes to secure the endpoint

* Returning Httpstatus as output and added exception handling changes

* Added Quadruple class to store the quadruples from the network in mongo
db

* [Q20-8] fixed package structure; added example network

* [Q20-8] fixed mapping object reference in example file

* [Q20-8] (hopefully) final data model and storage of networks

* [Q20-8] mapped events are stored in neo4j

+++ quadriga/src/main/java/edu/asu/diging/quadriga/core/model/elements/TermPart.java
@@ -6,13 +6,13 @@ public class TermPart extends Element {
     private String format;
     private String expression;
 
-    private Integer position;
+    private Long position;
 
     private VocabularyEntry normalization;
 
     private SourceReference source_reference;
 
-    public Integer getPosition() {
+    public Long getPosition() {
         return position;
     }
 
@@ -20,7 +20,7 @@ public class TermPart extends Element {
         return expression;
     }
 
-    public void setPosition(Integer position) {
+    public void setPosition(Long position) {
         this.position = position;
     }
